rename verification entity
