+++ portal-backend/sql/data-source.ts
@@ -9,7 +9,7 @@ import { TypeOrmRootModule } from '../src/app/configuration/TypeORM';
 import { Account } from '../src/users/entities/Account.entity';
 import { Session } from '../src/users/entities/Session.entity';
 import { User } from '../src/users/entities/User.entity';
-import { Verification } from '../src/users/entities/Verification';
+import { Verification } from '../src/users/entities/Verification.entity';
 
 @Module({
   imports: [EnvironmentConfigRootModule(), TypeOrmRootModule(true)],
