+++ portal-backend/test/test-helper/reset-db.ts
@@ -28,7 +28,7 @@ import { Subscription } from '../../src/subscription/entities/subscription.entit
 import { Account } from '../../src/users/entities/Account.entity';
 import { Session } from '../../src/users/entities/Session.entity';
 import { User } from '../../src/users/entities/User.entity';
-import { Verification } from '../../src/users/entities/Verification';
+import { Verification } from '../../src/users/entities/Verification.entity';
 
 // ---------------------------------------------------------------------------
 // Default users seeded after every reset
